requirements.txt
# FastAPI & ASGI server
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.9


# LangChain ecosystem
langchain
langchainhub
langchain-community
tiktoken==0.5.2

# Embeddings & vector DB
sentence-transformers==2.6.1
docx2txt
faiss-cpu==1.8.0

# File parsing & document handling
python-docx==1.1.0
pandas==2.2.2
lxml==5.2.1

# Utilities
pydantic
python-dotenv==1.0.1

# Optional Dev Tools
black==24.4.2
isort==5.13.2